feature: added try-except blocks; restructured code
